+++ templates/workspace_services/guacamole/user_resources/guacamole-azure-linuxvm/terraform/linuxvm.tf
@@ -44,6 +44,8 @@ resource "azurerm_linux_virtual_machine" "linuxvm" {
   disable_password_authentication = false
   admin_username                  = random_string.username.result
   admin_password                  = random_password.password.result
+  secure_boot_enabled             = local.secure_boot_enabled
+  vtpm_enabled                    = local.vtpm_enabled
 
   custom_data = data.template_cloudinit_config.config.rendered
 
@@ -72,7 +74,10 @@ resource "azurerm_linux_virtual_machine" "linuxvm" {
 
   tags = local.tre_user_resources_tags
 
-  lifecycle { ignore_changes = [tags] }
+  # ignore changes to secure_boot_enabled and vtpm_enabled as these are destructive
+  # (may be allowed once https://github.com/hashicorp/terraform-provider-azurerm/issues/25808 is fixed)
+  #
+  lifecycle { ignore_changes = [tags, secure_boot_enabled, vtpm_enabled] }
 }
 
 resource "azurerm_disk_encryption_set" "linuxvm_disk_encryption" {